refactor and add location weighting funtionallity

+++ TP5/src/TP5/TP5_funcs.h
@@ -4,6 +4,6 @@
 #include "../Utils/types.h"
 
 // TODO: implement the kmeans algorithm
-ppm_file kmeans(ppm_file image, int K, int init, int stop, int location);
+ppm_file kmeans(ppm_file image, int K, int init, int stop, double location);
 
 #endif
